Add in some enumerated flags, and do some tidying. (#8)
Enumerated flags are specified such that the flag value itself determines the value given to the property, by way of a map. For instance, mapping an enum, like so:
```
  val environment by parser.enumerated(
    "--staging" to Environment.STAGING,
    "--production" to Environment.PRODUCTION,
    help = "The deployment environment"
  )
```
When the user puts one of `--staging` or `--production` the appropriate enum value will be set.

Also updates the README and the examples a bit, and fleshes out some tests.